requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_pytorch_select=0.1=cpu_0
blas=1.0=mkl
ca-certificates=2020.7.22=0
certifi=2020.6.20=py38_0
cffi=1.14.3=py38hed5b41f_0
intel-openmp=2019.4=233
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libmklml=2019.0.5=0
llvm-openmp=10.0.0=h28b9765_0
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
ncurses=6.2=h0a44026_1
ninja=1.10.1=py38h879752b_0
numpy=1.19.1=py38h3b9f5b6_0
numpy-base=1.19.1=py38hcfb5961_0
openssl=1.1.1h=haf1e3a3_0
pip=20.2.2=py38_0
pycparser=2.20=py_2
python=3.8.5=h26836e1_1
pytorch=1.6.0=cpu_py38hd70000b_0
readline=8.0=h1de35cc_0
setuptools=49.6.0=py38_1
six=1.15.0=py_0
sqlite=3.33.0=hffcf06c_0
tk=8.6.10=hb0a8c7a_0
tqdm=4.49.0=py_0
wheel=0.35.1=py_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
